Domainsearch/requirements.txt
streamlit
python-dotenv
langchain
langchain-community
openai
google-search-results
